Use GetByBodyApi
